requirements.txt
python-telegram-bot==20.0 
google-generativeai==latest_version
python-dotenv==latest_version
requests

